analyze: refactor: separate expr-specific parts from general rewrite infrastructure
