--- c2rust-analyze/src/expr_rewrite/mod.rs
@@ -1,43 +1,0 @@
-use crate::context::{AnalysisCtxt, Assignment};
-use rustc_hir::BodyId;
-use rustc_middle::mir::Body;
-use rustc_middle::ty::TyCtxt;
-use rustc_span::Span;
-
-mod apply;
-mod hir_op;
-mod mir_op;
-mod span_index;
-
-pub use self::hir_op::Rewrite;
-
-pub fn gen_expr_rewrites<'tcx>(
-    acx: &AnalysisCtxt<'_, 'tcx>,
-    asn: &Assignment,
-    mir: &Body<'tcx>,
-    hir_body_id: BodyId,
-) -> Vec<(Span, Rewrite)> {
-    let mir_rewrites = mir_op::gen_mir_rewrites(acx, asn, mir);
-    let hir_rewrites = hir_op::gen_hir_rewrites(acx.tcx(), mir, hir_body_id, &mir_rewrites);
-    hir_rewrites
-}
-
-pub fn apply_rewrites(tcx: TyCtxt, rewrites: Vec<(Span, Rewrite)>) {
-    // TODO: emit new source code properly instead of just printing
-    let new_src = apply::apply_rewrites(tcx.sess.source_map(), rewrites);
-
-    for (filename, src) in new_src {
-        eprintln!("\n\n ===== BEGIN {:?} =====", filename);
-        for line in src.lines() {
-            // Omit filecheck directives from the debug output, as filecheck can get confused due
-            // to directives matching themselves (e.g. `// CHECK: foo` will match the `foo` in the
-            // line `// CHECK: foo`).
-            if let Some((pre, post)) = line.split_once("// CHECK") {
-                eprintln!("{}// (FileCheck directive omitted)", pre);
-            } else {
-                eprintln!("{}", line);
-            }
-        }
-        eprintln!(" ===== END {:?} =====", filename);
-    }
-}
