+++ c2rust-analyze/src/rewrite/span_index.rs
